requirements.txt
uvicorn
fastapi
pandas
tabula-py
python-multipart
